requirements.txt
absl-py==0.12.0
asgiref==3.3.1
astor==0.8.1
attrs==20.3.0
cached-property==1.5.2
cycler==0.10.0
Django==3.1.7
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.36.1
h5py==2.10.0
importlib-metadata==3.7.3
iniconfig==1.1.1
innvestigate==1.0.9
joblib==1.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
Pillow==8.1.2
plotly==4.14.3
pluggy==0.13.1
protobuf==3.15.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.59.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
